[coverage]: Modal.tsx, constants.ts
